Add background recording for forms

+++ collect_app/src/main/java/org/odk/collect/android/widgets/utilities/InternalRecordingRequester.java
@@ -1,45 +1,44 @@
 package org.odk.collect.android.widgets.utilities;
 
-import android.app.Activity;
-import android.util.Pair;
+import androidx.activity.ComponentActivity;
 
-import androidx.lifecycle.LifecycleOwner;
-
+import org.javarosa.core.model.FormIndex;
+import org.javarosa.core.model.data.StringData;
 import org.javarosa.form.api.FormEntryPrompt;
 import org.odk.collect.android.analytics.AnalyticsEvents;
+import org.odk.collect.android.application.Collect;
+import org.odk.collect.android.exception.JavaRosaException;
 import org.odk.collect.android.formentry.FormEntryViewModel;
+import org.odk.collect.android.formentry.FormIndexAnimationHandler;
+import org.odk.collect.android.formentry.saving.FormSaveViewModel;
 import org.odk.collect.android.listeners.PermissionListener;
-import org.odk.collect.android.utilities.FormEntryPromptUtils;
 import org.odk.collect.android.permissions.PermissionsProvider;
-import org.odk.collect.android.utilities.QuestionMediaManager;
+import org.odk.collect.android.utilities.FormEntryPromptUtils;
 import org.odk.collect.audiorecorder.recorder.Output;
 import org.odk.collect.audiorecorder.recording.AudioRecorderViewModel;
-
-import java.io.File;
-import java.util.function.Consumer;
+import org.odk.collect.audiorecorder.recording.RecordingSession;
 
 public class InternalRecordingRequester implements RecordingRequester {
 
-    private final Activity activity;
-    private final AudioRecorderViewModel viewModel;
+    private final ComponentActivity activity;
+    private final AudioRecorderViewModel audioRecorderViewModel;
     private final PermissionsProvider permissionsProvider;
-    private final LifecycleOwner lifecycleOwner;
-    private final QuestionMediaManager questionMediaManager;
     private final FormEntryViewModel formEntryViewModel;
+    private final FormSaveViewModel formSaveViewModel;
+    private final FormIndexAnimationHandler.Listener refreshListener;
 
-    public InternalRecordingRequester(Activity activity, AudioRecorderViewModel viewModel, PermissionsProvider permissionsProvider, LifecycleOwner lifecycleOwner, QuestionMediaManager questionMediaManager, FormEntryViewModel formEntryViewModel) {
+    public InternalRecordingRequester(ComponentActivity activity, AudioRecorderViewModel audioRecorderViewModel, PermissionsProvider permissionsProvider, FormEntryViewModel formEntryViewModel, FormSaveViewModel formSaveViewModel, FormIndexAnimationHandler.Listener refreshListener) {
         this.activity = activity;
-        this.viewModel = viewModel;
+        this.audioRecorderViewModel = audioRecorderViewModel;
         this.permissionsProvider = permissionsProvider;
-        this.lifecycleOwner = lifecycleOwner;
-        this.questionMediaManager = questionMediaManager;
         this.formEntryViewModel = formEntryViewModel;
-    }
+        this.formSaveViewModel = formSaveViewModel;
+        this.refreshListener = refreshListener;
 
-    @Override
-    public void onIsRecordingBlocked(Consumer<Boolean> isRecordingBlockedListener) {
-        viewModel.getCurrentSession().observe(lifecycleOwner, session -> {
-            isRecordingBlockedListener.accept(session != null && session.getFile() == null);
+        audioRecorderViewModel.getCurrentSession().observe(activity, session -> {
+            if (session != null && session.getFile() != null) {
+                handleRecording(session);
+            }
         });
     }
 
@@ -50,11 +49,11 @@ public class InternalRecordingRequester implements RecordingRequester {
             public void granted() {
                 String quality = FormEntryPromptUtils.getAttributeValue(prompt, "quality");
                 if (quality != null && quality.equals("voice-only")) {
-                    viewModel.start(prompt.getIndex().toString(), Output.AMR);
+                    audioRecorderViewModel.start(prompt.getIndex(), Output.AMR);
                 } else if (quality != null && quality.equals("low")) {
-                    viewModel.start(prompt.getIndex().toString(), Output.AAC_LOW);
+                    audioRecorderViewModel.start(prompt.getIndex(), Output.AAC_LOW);
                 } else {
-                    viewModel.start(prompt.getIndex().toString(), Output.AAC);
+                    audioRecorderViewModel.start(prompt.getIndex(), Output.AAC);
                 }
             }
 
@@ -67,29 +66,25 @@ public class InternalRecordingRequester implements RecordingRequester {
         formEntryViewModel.logFormEvent(AnalyticsEvents.AUDIO_RECORDING_INTERNAL);
     }
 
-    @Override
-    public void onRecordingInProgress(FormEntryPrompt prompt, Consumer<Pair<Long, Integer>> durationListener) {
-        viewModel.getCurrentSession().observe(lifecycleOwner, session -> {
-            if (session != null && session.getId().equals(prompt.getIndex().toString()) && session.getFailedToStart() == null) {
-                durationListener.accept(new Pair<>(session.getDuration(), session.getAmplitude()));
-            }
-        });
-    }
+    private void handleRecording(RecordingSession session) {
+        formSaveViewModel.createAnswerFile(session.getFile()).observe(activity, result -> {
+            if (result != null && result.isSuccess()) {
+                audioRecorderViewModel.cleanUp();
 
-    @Override
-    public void onRecordingFinished(FormEntryPrompt prompt, Consumer<File> recordingAvailableListener) {
-        viewModel.getCurrentSession().observe(lifecycleOwner, session -> {
-            if (session != null && session.getId().equals(prompt.getIndex().toString()) && session.getFile() != null) {
-                questionMediaManager.createAnswerFile(session.getFile()).observe(lifecycleOwner, result -> {
-                    if (result != null) {
-                        if (result.isSuccess()) {
-                            session.getFile().delete();
-                        }
-
-                        viewModel.cleanUp();
-                        recordingAvailableListener.accept(result.getOrNull());
+                try {
+                    if (session.getId() instanceof FormIndex) {
+                        FormIndex formIndex = (FormIndex) session.getId();
+                        formSaveViewModel.replaceAnswerFile(formIndex.toString(), result.getOrNull().getAbsolutePath());
+                        Collect.getInstance().getFormController().answerQuestion(formIndex, new StringData(result.getOrNull().getName()));
+                        refreshListener.onScreenRefresh();
                     }
-                });
+                } catch (JavaRosaException e) {
+                    // ?
+                }
+
+                if (!session.getId().equals("background")) {
+                    session.getFile().delete();
+                }
             }
         });
     }
